+++ bindings/nostr-sdk-ffi/python/examples/whitelist.py
@@ -1,5 +1,5 @@
 import asyncio
-from nostr_sdk import PublicKey, Client, Filter, Kind, init_logger, LogLevel, Options, RelayFilteringMode
+from nostr_sdk import PublicKey, ClientBuilder, Filter, Kind, init_logger, LogLevel, Options, RelayFilteringMode
 from datetime import timedelta
 
 
@@ -24,7 +24,7 @@ async def main():
     # Get events
     f = Filter().authors([whitelisted_public_key, not_whitelisted_public_key]).kind(Kind(0))
     events = await client.fetch_events([f], timedelta(seconds=10))
-    print(f"Received {events.__len__()} events")
+    print(f"Received {events.len()} events")
 
 
 if __name__ == '__main__':