+++ bindings/nostr-sdk-ffi/src/protocol/nips/nip17.rs
@@ -5,7 +5,7 @@ use std::ops::Deref;
 use std::sync::Arc;
 
 use crate::error::Result;
-use crate::protocol::signer::{NostrSigner, NostrSignerFFI2Rust};
+use crate::protocol::signer::NostrSigner;
 use crate::protocol::{Event, PublicKey, Tag};
 
 /// Private Direct message
@@ -13,14 +13,13 @@ use crate::protocol::{Event, PublicKey, Tag};
 /// <https://github.com/nostr-protocol/nips/blob/master/17.md>
 #[uniffi::export(async_runtime = "tokio", default(rumor_extra_tags = []))]
 pub async fn make_private_msg(
-    signer: Arc<dyn NostrSigner>,
+    signer: &NostrSigner,
     receiver: &PublicKey,
     message: &str,
     rumor_extra_tags: Vec<Arc<Tag>>,
 ) -> Result<Event> {
-    let signer = NostrSignerFFI2Rust::new(signer);
     Ok(nostr::EventBuilder::private_msg(
-        &signer,
+        signer.deref(),
         **receiver,
         message,
         rumor_extra_tags